requirements.txt
pyjwt==2.10.1
supabase==2.27.0

# ----------------------------
# Core FastAPI stack
# ----------------------------
fastapi==0.128.0
uvicorn[standard]==0.35.0

# ----------------------------
# Database (PostgreSQL)
# ----------------------------
sqlalchemy==2.0.43
alembic==1.17.2
psycopg2-binary==2.9.9
asyncpg==0.30.0

# ----------------------------
# Pydantic & settings
# ----------------------------
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0

# ----------------------------
# Authentication & Security
# ----------------------------
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20

# ----------------------------
# Environment
# ----------------------------
python-dotenv==1.1.1
